refactored DriverIntegrationTests and added comments
